Refine token input prompts for GitLab API
Improved the user experience in the setup process for the GitLab API by adding a validation function ('no_whitespace') to the input prompts for user token and token name. The function checks if the inputs are either empty or only contains whitespaces, triggering an error if so. This prevents empty or whitespace-only entries from being accepted, which were previously causing errors in the setup process. Also welcome prompt was added to notify user how to obtain gitlab API token.